[INJIMOB-3539] refactor: move CredentialInputDescriptorMapping data class
