+++ kotlin/openID4VP/src/commonMain/kotlin/io/mosip/openID4VP/authorizationResponse/CredentialInputDescriptorMapping.kt
@@ -1,4 +1,4 @@
-package io.mosip.openID4VP.authorizationResponse.mapping
+package io.mosip.openID4VP.authorizationResponse
 
 import io.mosip.openID4VP.constants.FormatType
