+++ kotlin/openID4VP/src/commonMain/kotlin/io/mosip/openID4VP/authorizationResponse/unsignedVPToken/UnsignedVPTokenBuilder.kt
@@ -1,6 +1,6 @@
 package io.mosip.openID4VP.authorizationResponse.unsignedVPToken
 
-import io.mosip.openID4VP.authorizationResponse.mapping.CredentialInputDescriptorMapping
+import io.mosip.openID4VP.authorizationResponse.CredentialInputDescriptorMapping
 
 internal interface UnsignedVPTokenBuilder {
     fun build(credentialInputDescriptorMappings : List<CredentialInputDescriptorMapping>) : Pair<Any?, UnsignedVPToken>
