+++ kotlin/openID4VP/src/jvmMain/kotlin/io/mosip/openID4VP/common/Encoder.kt
@@ -1,4 +1,4 @@
-package io.mosip.openID4VP.common
+  package io.mosip.openID4VP.common
 
 import java.util.Base64.getUrlEncoder
